requirements.txt
asgiref==3.8.1
bitarray==2.9.2
Django==4.2
pycryptodome==3.20.0
sqlparse==0.5.0
tzdata==2024.1
